simplify code
move multiple if ($opt_akotfps) branches and calls to add_akotfps_data
into the generate_font_snippet function.

+++ cjk-gs-integrate.pl
@@ -500,12 +500,8 @@ sub do_otf_fonts {
     if $opt_texmflink;
   for my $k (keys %fontdb) {
     if ($fontdb{$k}{'available'} && $fontdb{$k}{'type'} eq 'OTF') {
-      if ($opt_akotfps) {
-        add_akotfps_data($k);
-      } else {
-        generate_font_snippet($fontdest,
-          $k, $fontdb{$k}{'class'}, $fontdb{$k}{'target'});
-      }
+      generate_font_snippet($fontdest,
+        $k, $fontdb{$k}{'class'}, $fontdb{$k}{'target'});
       link_font($fontdb{$k}{'target'}, $ciddest, $k);
       link_font($fontdb{$k}{'target'}, "$opt_texmflink/$otf_pathpart", "$fontdb{$k}{'origname'}.otf")
         if $opt_texmflink;
@@ -516,6 +512,10 @@ sub do_otf_fonts {
 sub generate_font_snippet {
   my ($fd, $n, $c, $f) = @_;
   return if $dry_run;
+  if ($opt_akotfps) {
+    add_akotfps_data($n);
+    return;
+  }
   for my $enc (@{$encode_list{$c}}) {
     if ($opt_remove) {
       unlink "$fd/$n-$enc" if (-f "$fd/$n-$enc");
@@ -638,23 +638,15 @@ sub do_nonotf_fonts {
     if $opt_texmflink;
   for my $k (keys %fontdb) {
     if ($fontdb{$k}{'available'} && $fontdb{$k}{'type'} eq 'TTF') {
-      if ($opt_akotfps) {
-        add_akotfps_data($k);
-      } else {
-        generate_font_snippet($fontdest,
-          $k, $fontdb{$k}{'class'}, $fontdb{$k}{'target'});
-      }
+      generate_font_snippet($fontdest,
+        $k, $fontdb{$k}{'class'}, $fontdb{$k}{'target'});
       $outp .= generate_cidfmap_entry($k, $fontdb{$k}{'class'}, $fontdb{$k}{'ttfname'}, $fontdb{$k}{'subfont'});
       link_font($fontdb{$k}{'target'}, $cidfsubst, $fontdb{$k}{'ttfname'});
       link_font($fontdb{$k}{'target'}, "$opt_texmflink/$ttf_pathpart", $fontdb{$k}{'ttfname'})
         if $opt_texmflink;
     } elsif ($fontdb{$k}{'available'} && $fontdb{$k}{'type'} eq 'TTC') {
-      if ($opt_akotfps) {
-        add_akotfps_data($k);
-      } else {
-        generate_font_snippet($fontdest,
-          $k, $fontdb{$k}{'class'}, $fontdb{$k}{'target'});
-      }
+      generate_font_snippet($fontdest,
+        $k, $fontdb{$k}{'class'}, $fontdb{$k}{'target'});
       $outp .= generate_cidfmap_entry($k, $fontdb{$k}{'class'}, $fontdb{$k}{'ttcname'}, $fontdb{$k}{'subfont'});
       link_font($fontdb{$k}{'target'}, $cidfsubst, $fontdb{$k}{'ttcname'});
       link_font($fontdb{$k}{'target'}, "$opt_texmflink/$ttf_pathpart", $fontdb{$k}{'ttcname'})
@@ -662,12 +654,8 @@ sub do_nonotf_fonts {
     } elsif ($fontdb{$k}{'available'} && $fontdb{$k}{'type'} eq 'OTC') {
     # currently ghostscript does not have OTC support; not creating gs resource
     print_ddebug("gs does not support OTC, not creating gs resource for $k\n");
-    # if ($opt_akotfps) {
-    #   add_akotfps_data($k);
-    # } else {
-    #   generate_font_snippet($fontdest,
-    #     $k, $fontdb{$k}{'class'}, $fontdb{$k}{'target'});
-    # }
+    # generate_font_snippet($fontdest,
+    #  $k, $fontdb{$k}{'class'}, $fontdb{$k}{'target'});
     # $outp .= generate_cidfmap_entry($k, $fontdb{$k}{'class'}, $fontdb{$k}{'otcname'}, $fontdb{$k}{'subfont'});
     # link_font($fontdb{$k}{'target'}, $cidfsubst, $fontdb{$k}{'otcname'});
       link_font($fontdb{$k}{'target'}, "$opt_texmflink/$otf_pathpart", $fontdb{$k}{'otcname'})
@@ -736,11 +724,7 @@ sub do_aliases {
     }
     # we also need to create font snippets in Font (or add configuration)
     # for the aliases!
-    if ($opt_akotfps) {
-      add_akotfps_data($al);
-    } else {
-      generate_font_snippet($fontdest, $al, $class, $target);
-    }
+    generate_font_snippet($fontdest, $al, $class, $target);
     if ($class eq 'Japan') {
       push @jal, "/$al /$target ;";
     } elsif ($class eq 'Korea') {
